Start converting types to use pointer fields.
As we venture into the world of PATCH-type update requests, it becomes
obvious very quickly that a simple empty/non-empty check of any given
type is not sufficient. We need a tri-value for each type, which allows
clients to specify a value (including the empty value), or no value at
all, which is useful to indicate "no change".