Do not use pointers for the option parameters
The `JSONAPI` spec requires that you always send a valid `JSONAPI` document when doing a `PATCH` or `POST`, even when you don’t send any data (attributes).

So either we have to do:

```go
if options == nil {
  options = &CreateXXXOptions{}
}
```

In every function where we accept an options struct, or we can just require the user to actualy send a stuct and not give the option to send a `nil` pointer.

I’m not 100% happy with this solution for some reason, but I think this is the cleanest way with the least amount of code that will always give the desired result.

And since all the option structs are very small (they have just a few fields) and all the fields are pointers, the overhead of passing the struct instead of a pointer to a struct seems very small in this case.