Check for STX and ETX when polling for logs
By checking for the STX and ETX characters, we have a way to determine if the process generating the logs is finished.

This prevents the need to poll for the status of a plan or apply during its execution.

The STX and ETX ASCII contol characters are part of the transmission control characters which are intended to structure a data stream.

The Start of Text character (STX) is to mark the start of the textual part of a stream and the End of Text character (ETX) marked the end of the data of a message.