Client.RemoteAPIVersion: get the remote API version number
Newer versions of Terraform Cloud and Enterprise return the API version in
an HTTP header on all API responses.

This client implementation already always makes a pre-flight request to
a special "ping" endpoint as part of its initialization anyway, previously
to configure the rate limiting settings. To avoid additional requests and
keep this relatively simple, we'll now also use that request to populate
the remote API version number and thus make it available immediately to
callers via the RemoteAPIVersion method.