Revert Enforcement type (#211)
In #197 we changed the type of Policy.Enforce from []*Enforcement to []Enforcement. This had downstream impact on the terraform-provider, which relies on this client (as well as any other 3rd parties that use this client!)

This PR primarily reverts the type changes of Policy.Enforce back to being a []*Enforcement. It also updates the hashicorp/jsonapi package to reference the latest version.