Backport: Revert Enforcement type (#211)
This commit is for a separate backported release of the following fix,
for the 0.13.x series of the client.

In #197 we changed the type of Policy.Enforce from []*Enforcement to []Enforcement. This had downstream impact on the terraform-provider, which relies on this client (as well as any other 3rd parties that use this client!)

This PR primarily reverts the type changes of Policy.Enforce back to being a []*Enforcement. It also updates the hashicorp/jsonapi package to reference the latest version.

It also includes a field tag change necessary from #209