Add Upload() for registry module versions
Similar to configuration versions, with the slight modification that I
don't believe client users should care about piping the correct upload
link to Upload(), or it's little more than syntatic sugar for a simple
PUT request.

With this, just provide a newly created RMV to Upload() with a path to
your module files and away you go.