Correct RunStatusTimestamps (#227)
These timestamp fields are wrong, and were wrong when originally
introduced pre-Terraform Cloud. As they've never been populated
correctly before, no client could be relying on them and the ones that
don't work are removed as a 'not-so-breaking' change, with all the
modern timestamps on runs added.

This change does not include tests as to do so would require queuing a
run for every permutation of states (extremely expensive) only to test
that unmarshaling is done properly in the jsonapi package (done
elsewhere).