running script to generate mocks inside of a github act… (#263)
* Add GitHub action workflow to verify that mocks have been generated after a change in interfaces

The purpose of this workflow is to make sure that our auto generated mocks are up to date. If a contributor makes a change to any go interfaces and has forgotten to run the script to generate mocks for the new interface implementation, the workflow will fail. If there are no changes to the interfaces, then the workflow will succeed.
The workflow is implemented in multiple steps such as setting up a go environment using the go version defined in .go-version file, installing mockgen latest version, running the script generate_mocks.sh and looking for a diff in the repo after running the script. If there is a diff, then the exit code will be 1 and the workflow will fail.