Added a convenience function for attaching run tasks to a workspace
WorkspaceRunTasks.Create() is not as intuitive to the user and doesn't clearly express the action
that's happening. When a workspace run task is created the action is "attaching a run task to a
workspace". The API docs also state this action instead of "create workspace run task". Therefore
this method adds a more expressive wrapper over Create() keeping the user within the "Run Tasks" context.