fix: API does not allow ssh key text value updating
From the very beginning it's not been possible to update the value of an SSH key text. Indeed, the tests didn't even test that the value was updated!

The reason for this is the non-repudiation of security auditing. For example, if an attacker uses the ssh key to access a resource, the security audit of the ssh authentication can be trusted to mean key X was always key X and never updated.