Include field string to an slice of typed strings (#294)
Many of our APIs offer an "include" capability to include other resources. One can include many other resources for the GET endpoints. In the go-tfe client, we just set this as a string for the Include Param in ReadOptions and ListOptions. This works, but its is not entirely useful as it requires someone using this client to then reference the API to see what are the strings available. Occasionally, they made even have a typo in there and not notice until a little late in their process, when they get the Atlas error for passing an incorrect include param. It would make for a better client experience to know from the Go-tfe code what resources are possible.
This is doable via the use of a typed string in Go.