Return error message when a workspace is locked by a pending run
The current implementation of this client would default to "workspaces already unlocked"
whenever a status 409 is received from the workspace unlock endpoint. This however paints half the
picture since the endpoint can also return a 409 Conflict error whenever a workspace is locked by a run. The
current implementation for the method responsible for returning the correct error message doesn't parse the error payload
when handling status 409.

Parsing the error payload was then moved to it's own method, so we can determine the nature of the error when handling 409s.

A simple matcher function was written to see if the error detail contains a "locked by Run" substring which indicates
the workspace was locked by the run and thus return the correct error message