Use consistent naming across resource methods
This commit introduces a refactor for several methods that express an action on a relationship between two resources. By using
verbs, we can be more specific as to what action we're performing when calling these methods. An example would be
ModuleConsumers() -- from this method signature we can't necessarily tell what action this method is performing without
referring to the docs. Renaming it to ListModuleConsumers() we can now assess what the function does more conveniently.