Add `integration` tag to lintFlags in vscode settings
VSCode's integration with the staticcheck linter was marking all our test helper
functions as unused code, which is untrue! The linter doesn't seem to inherit
the `buildFlags`, and has its own flags setting. With this change, it properly
notices that our test helpers are actually important.

The linter _does_ seem to inherit the `go.buildTags` setting, so that might be a
more complete way to address this. But I wasn't sure yet if that would have any
wider effects we don't want.