Use require to prevent a panic when dereferencing a nil pointer
This is a workaround to simply prevent a panic from occurring. What's happening here
is the test is expecting two relationships to be included in the response, CurrentVersion and NewestVersion, but was
subject to the overall flakiness of our CI environment. This commit does not fix nor prevent the flakiness.

Why `require` and not `assert`? `require` makes the same calls as `assert` except it will terminate the execution
of a test upon failure. Using `assert` would still continue the test's execution, dereference a nil pointer, and boom panic!