Hopefully fix flaky admin workspace list tests
One of these tests was flaking on this PR, probably because it was tacitly
expecting that fewer than 20 workspaces ever exist at once on the CI TFC
instance. The more tests we add that create temporary workspaces (like the ones
in this PR), the more likely that test was to flake.

Testing shared global state is always nasty and error-prone, but this commit
adjusts a couple tests to make less precise assertions that hopefully still test
the behavior of interest.