Response header hooks
The API for this client library is focused on returning the primary data
associated with each API response, which typically means just the body
of the response.

Sometimes clients will also need to react to cross-cutting metadata such
as expiration times, cache control guidance, and rate limiting
information, which isn't a direct part of the data being requested but
can nonetheless affect the behavior of the client. This information is
typically returned in HTTP response header fields.

To give access to this information without a breaking change to the API,
this uses the context.Context API to allow the rare caller that needs it
to register a hook through which it will be notified about the response
header in any case where the request succeeded enough for there to be one.

Most clients will not need this facility, which justifies the light abuse
of the context.Context API for passing in this optional hook, even though
this isn't the sort of cross-cutting concern context.Context should
typically be used for.