Await async state version processing in TestStateVersionsUpload
TFC does some async processing on uploaded state versions; they're not
guaranteed to have `finalized` status by the time the upload call is
acknowledged. That means this test flakes. (Though it wasn't flaking in CI until
today, since it was previously being skipped as beta.)

This adds a state version clone of the run polling function. Yeah, it's
copypasta code, but it seemed premature to try and come up with some kind of
ReadWithStatus interface that could cope with all the distinct string types etc.
Possibly worth revisiting later.