Fix validation check for "cost_estimated" [TF-24176]
The validation check should have RunCostEstimated, a RunStatus defined as "cost_estimated".
RunCostEstimate is a RunIncludeOpt.

I added a unit test for validateAdminRunFilterParams, and found that some of the RunStatus values are not accepted.
I've moved those to a separate check for invalid values, but possibly validateAdminRunFilterParams should be changed to accept them too?
If so I can update this PR.

Note the ticket TF-24176 refers to https://pkg.go.dev/github.com/hashicorp/go-tfe#RunStatus, which does list all the RunStatus strings, so customers are under the impression they can all be used as filters.